refactored and fixed functions and sql CAF rules
